core/core/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

azure-ai-formrecognizer==3.3.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
langchain==0.1.12
langchain-community==0.0.28
langchain-openai==0.0.8
openai==1.14.0
azure-search-documents==11.4.0
tiktoken==0.6.0
puremagic==1.15
pillow==10.2.0

dataclasses-json==0.6.3
azure-messaging-webpubsubservice==1.0.1
pyodbc==5.0.1
SQLAlchemy==2.0.23
